pyproject.toml
[project]
name = "django-tasks-gcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Chris Muthig", email = "camuthig@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "django>=6.0a1,<7.0",
    "google-cloud-tasks>=2.19.3",
]

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
